requirements.txt
python-dotenv>=1.0.0
pdfplumber>=0.10.0
google-generativeai>=0.3.0
python-slugify>=8.0.0
weasyprint>=60.0
SQLAlchemy>=2.0.0
pydantic-ai[logfire]>=0.1.6
jinja2>=3.1.0
pytest>=8.1.1
beautifulsoup4>=4.12.0
google-cloud-storage>=2.14.0
slack-sdk>=3.23.0
google-cloud-secret-manager
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.7
pydantic>=2.6.0
pydantic-ai>=1.0.0

-e .

# Additional dependencies not required for package installation
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
mypy>=1.0.0